moved forms components into forms instead of feeback
