+++ frontend/src/assets/scss/bleu_vert.scss
@@ -43,6 +43,9 @@
 .p-g-70 {
 	padding:0 0 0 70px;
 }
+.m-l-15 {
+    padding:0 15px 0 0;
+}
 .p-spec{
 	padding-left:0px;
 	padding-right:0px;
