+++ frontend/src/router/index.js
@@ -6,8 +6,8 @@ import Search from '@/components/Search'
 import Report from '@/components/Report'
 import Faq from '@/components/infos/Faq'
 import Legal from '@/components/infos/Legal'
-import Contact from '@/components/feedback/Contact'
-import Feedback from '@/components/feedback/Feedback'
+import Contact from '@/components/forms/Contact'
+import Feedback from '@/components/forms/Feedback'
 import Buyer from '@/components/infos/Buyer'
 import NotFound from '@/components/infos/NotFound'
